Refactor: separate helpers
