+++ application/src/DataSources/References/CWE.ts
@@ -2,7 +2,7 @@ import {DOMParser} from '@xmldom/xmldom';
 import XmlReader from 'xml-reader';
 import XmlQuery from 'xml-query';
 import {select} from 'xpath';
-import {tidyString} from '../../Helpers';
+import {tidyString} from '../../Helpers/Strings';
 
 import CWEDataset from '../../assets/cwe.xml';
