+++ application/src/Helpers/Strings.test.ts
@@ -1,0 +1,19 @@
+import {isURL} from './Strings';
+
+describe('isURL', () => {
+  test('returns true for an http url', () => {
+    expect(isURL('https://example.com')).toBeTruthy();
+  });
+
+  test('returns false for a unix filesystem path', () => {
+    expect(isURL('/path/to/file')).toBeFalsy();
+  });
+
+  test('returns false for a windows filesystem path', () => {
+    expect(isURL('C:\\Path\\To\\File')).toBeFalsy();
+  });
+
+  test('returns false for null input', () => {
+    expect(isURL(undefined)).toBeFalsy();
+  });
+});
